refactor(tui): move file link parsing into component
